% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SecondaryDataFunctions.R
\name{ExtractExperimentGuess}
\alias{ExtractExperimentGuess}
\title{This function extracts the answer to the post experiment questinnnaire.
That is, wether the participant correct guessed the purpose of the experiment.
The input data is the finger counting data structure.}
\usage{
ExtractExperimentGuess(data)
}
\description{
This function extracts the answer to the post experiment questinnnaire.
That is, wether the participant correct guessed the purpose of the experiment.
The input data is the finger counting data structure.
}
