% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeTrackerFunctions.R
\name{ExtractEyetracker}
\alias{ExtractEyetracker}
\title{This function is used for extracting the eye tracker data that is written
into the matlab struct from the main PTB script.
It produces an output that includes whether to accept or reject the trail
and a number corresponding the max horizontal deviation in degrees
trials are marked as rejected in the max horizontal deviation exceeds
2 degrees. Deviations before the digit appeared are not marked.
These parameters can be over-written by editing the function directly
as they have been hard-coded for simiplicity}
\usage{
ExtractEyetracker(this.SubjectData)
}
\description{
If no eye tracking data exists, then the output shows accept for all trials
and lists the max horizontal deviation as zero
}
