% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FingerCountingFunctions.R
\name{ExtractFingerData}
\alias{ExtractFingerData}
\title{This function extracts the results of the finger counting task.
It produces an output that include the first finger used in for each
of the sentences. It also marks the version. This corresponds to whether
the experimenter recorded all the fingers used (old) or whether they simply
recorded the first finger used on each hand}
\usage{
ExtractFingerData(fingerData)
}
\description{
This function extracts the results of the finger counting task.
It produces an output that include the first finger used in for each
of the sentences. It also marks the version. This corresponds to whether
the experimenter recorded all the fingers used (old) or whether they simply
recorded the first finger used on each hand
}
