% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResponseFunctions.R
\name{ExtractResponse}
\alias{ExtractResponse}
\title{This function extracts the reaction time data.
It produces an output for each trial, including the digit shown the target
location, the delay duration, and whether a response was made}
\usage{
ExtractResponse(this.SubjectData)
}
\description{
This function extracts the reaction time data.
It produces an output for each trial, including the digit shown the target
location, the delay duration, and whether a response was made
}
