% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SecondaryDataFunctions.R
\name{ExtractSecondaryData}
\alias{ExtractSecondaryData}
\title{This function extracts the secondary data
It produces an output that contains the handedness score, the maths test score,
the language response code (1,2,3), the age entered, and the aAMAS
(math anxiety score)}
\usage{
ExtractSecondaryData(this.SubjectData)
}
\description{
This function extracts the secondary data
It produces an output that contains the handedness score, the maths test score,
the language response code (1,2,3), the age entered, and the aAMAS
(math anxiety score)
}
