% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{FixExcludeData}
\alias{FixExcludeData}
\title{This function is applies a fix to the Exclude Data column in a lab's
FileList file. This is needed because different labs took different
approaches to filling in this column. Some examples, include:
Using True/False, using YES/NO, using 1/0, only making FALSE etc
This rewrites the column to a start of TRUE/FALSE}
\usage{
FixExcludeData(this.FileList)
}
\description{
This function is applies a fix to the Exclude Data column in a lab's
FileList file. This is needed because different labs took different
approaches to filling in this column. Some examples, include:
Using True/False, using YES/NO, using 1/0, only making FALSE etc
This rewrites the column to a start of TRUE/FALSE
}
