% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GetAllNodeFiles}
\alias{GetAllNodeFiles}
\title{GetAllNodeFile downloads all the files from a particular node. It stores the
files in subfolder /data/{nodename} in the current working folder. This only
downloads .xls and .mat files, and does not download the .edf files which are
not needed for the analysis}
\usage{
GetAllNodeFiles(node)
}
\description{
GetAllNodeFile downloads all the files from a particular node. It stores the
files in subfolder /data/{nodename} in the current working folder. This only
downloads .xls and .mat files, and does not download the .edf files which are
not needed for the analysis
}
