% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GetLabNames}
\alias{GetLabNames}
\title{Generates two vectors, one called LabNames and one called Nodes
These contain the lab names and node codes for the replication project}
\usage{
GetLabNames()
}
\description{
Generates two vectors, one called LabNames and one called Nodes
These contain the lab names and node codes for the replication project
}
