% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GetNodeFiles}
\alias{GetNodeFiles}
\title{Gets all the files that belong to a particular node}
\usage{
GetNodeFiles(dataFolder, this.node)
}
\description{
Gets all the files that belong to a particular node
}
