% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ReadMatFile}
\alias{ReadMatFile}
\title{ReadMatFile is a wrapper function for the readMat from the R.matlab package.
It's adds basic error checking. This is implemented by checking that it is a
MATLAB 5.0 MAT file. All files generated by the Fischer replication should be
of this format. If the file type is incorrect then a NULL is returned}
\usage{
ReadMatFile(filename)
}
\description{
ReadMatFile is a wrapper function for the readMat from the R.matlab package.
It's adds basic error checking. This is implemented by checking that it is a
MATLAB 5.0 MAT file. All files generated by the Fischer replication should be
of this format. If the file type is incorrect then a NULL is returned
}
